+++ telegram/__init__.py
@@ -20,6 +20,7 @@
 
 from .base import TelegramObject
 from .user import User
+from .files.chatphoto import ChatPhoto
 from .chat import Chat
 from .chatmember import ChatMember
 from .files.photosize import PhotoSize
@@ -119,5 +120,5 @@ __all__ = [
     'MAX_FILESIZE_DOWNLOAD', 'MAX_FILESIZE_UPLOAD', 'MAX_MESSAGES_PER_SECOND_PER_CHAT',
     'MAX_MESSAGES_PER_SECOND', 'MAX_MESSAGES_PER_MINUTE_PER_GROUP', 'WebhookInfo', 'Animation',
     'Game', 'GameHighScore', 'VideoNote', 'LabeledPrice', 'SuccessfulPayment', 'ShippingOption',
-    'ShippingAddress', 'PreCheckoutQuery', 'OrderInfo', 'Invoice', 'ShippingQuery'
+    'ShippingAddress', 'PreCheckoutQuery', 'OrderInfo', 'Invoice', 'ShippingQuery', 'ChatPhoto'
 ]