+++ telegram/chat.py
@@ -19,7 +19,7 @@
 # along with this program.  If not, see [http://www.gnu.org/licenses/].
 """This module contains an object that represents a Telegram Chat."""
 
-from telegram import TelegramObject
+from telegram import TelegramObject, ChatPhoto
 
 
 class Chat(TelegramObject):
@@ -38,6 +38,11 @@ class Chat(TelegramObject):
         last_name (str): Optional. Last name of the other party in a private chat.
         all_members_are_administrators (bool): Optional. True if a group has `All Members Are
                 Admins` enabled.
+        photo (:class:`telegram.ChatPhoto`): Optional. Chat photo. Returned only in getChat.
+        description	(str): Optional. Description, for supergroups and channel chats. Returned
+                only in getChat.
+        invite_link (str): Optional. Chat invite link, for supergroups and channel chats. Returned
+                only in getChat.
 
         CHANNEL (str): 'channel'
         GROUP (str): 'group'
@@ -57,6 +62,11 @@ class Chat(TelegramObject):
         last_name(Optional[str]): Last name of the other party in a private chat.
         all_members_are_administrators (Optional[bool]): Optional. True if a group has `All Members
                 Are Admins` enabled.
+        photo (Optional[:class:`telegram.ChatPhoto`]): Chat photo. Returned only in getChat.
+        description	(Optional[str]): Description, for supergroups and channel chats. Returned
+                only in getChat.
+        invite_link (Optional[str]): Chat invite link, for supergroups and channel chats. Returned
+                only in getChat.
         bot (Optional[:class:`telegram.Bot`]): The Bot to use for instance methods
         **kwargs (dict): Arbitrary keyword arguments.
 
@@ -76,6 +86,9 @@ class Chat(TelegramObject):
                  last_name=None,
                  all_members_are_administrators=None,
                  bot=None,
+                 photo=None,
+                 description=None,
+                 invite_link=None,
                  **kwargs):
         # Required
         self.id = int(id)
@@ -86,6 +99,9 @@ class Chat(TelegramObject):
         self.first_name = first_name
         self.last_name = last_name
         self.all_members_are_administrators = all_members_are_administrators
+        self.photo = photo
+        self.description = description
+        self.invite_link = invite_link
 
         self.bot = bot
         self._id_attrs = (self.id,)
@@ -103,6 +119,8 @@ class Chat(TelegramObject):
 
         if not data:
             return None
+
+        data['photo'] = ChatPhoto.de_json(data.get('photo'), bot)
 
         return Chat(bot=bot, **data)