+++ telegram/utils/helpers.py
@@ -19,14 +19,56 @@
 """ This module contains helper functions """
 
 import re
+from datetime import datetime
 
 try:
     from html import escape as escape_html  # noqa: F401
 except ImportError:
     from cgi import escape as escape_html  # noqa: F401
 
+# Not using future.backports.datetime here as datetime value might be an input from the user,
+# making every isinstace() call more delicate. So we just use our own compat layer.
+if hasattr(datetime, 'timestamp'):
+    # Python 3.3+
+    def _timestamp(dt_obj):
+        return dt_obj.timestamp()
+else:
+    # Python < 3.3 (incl 2.7)
+    from time import mktime
+
+    def _timestamp(dt_obj):
+        return mktime(dt_obj.timetuple())
+
 
 def escape_markdown(text):
     """Helper function to escape telegram markup symbols"""
     escape_chars = '\*_`\['
     return re.sub(r'([%s])' % escape_chars, r'\\\1', text)
+
+
+def to_timestamp(dt_obj):
+    """
+    Args:
+        dt_obj (:class:`datetime.datetime`):
+
+    Returns:
+        int:
+    """
+    if not dt_obj:
+        return None
+
+    return int(_timestamp(dt_obj))
+
+
+def from_timestamp(unixtime):
+    """
+    Args:
+        unixtime (int):
+
+    Returns:
+        datetime.datetime:
+    """
+    if not unixtime:
+        return None
+
+    return datetime.fromtimestamp(unixtime)